Move existing annotations to new annotation directory.
